Add Users Roles and configure login page
On ApplicationDbContext we override database table names with new ones.
Deleted old database and create new Migrations then update database.
We added new functionality to login page. Now we able to login with both
username and email. After that for able to assigning roles to users,
first created Enums/Roles.cs then seeded this roles IdentityRole.
Lastly we seeds required roles in Program.cs if that is not already
seeded.